server/requirements.txt
aniso8601==9.0.1
anyio==3.5.0
ariadne==0.14.1
asgiref==3.5.0
autopep8==1.6.0
click==8.0.4
graphene==3.0
graphql-core==3.1.7
graphql-relay==3.1.5
h11==0.13.0
idna==3.3
neo4j-driver==4.3.6
neobolt==1.7.17
neomodel==4.0.8
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
Shapely==1.7.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.1.1
uvicorn==0.17.5
